+++ app/src/main/java/db/entities/CardStack.kt
@@ -7,7 +7,7 @@ import java.io.Serializable
 
 @Entity
 data class CardStack(
-    @PrimaryKey val cardStackId: Int,
+    @PrimaryKey val cardStackId: Long,
     @ColumnInfo(name = "courseId") val courseId: Long, //Foreign key to course
 //    @ColumnInfo(name = "active") val active: Boolean,
 //    @ColumnInfo(name = "cards") val cards: List<Card>,
